+++ internal/staticapp/generate.go
@@ -1,0 +1,3 @@
+package staticapp
+
+//go:generate go run generate_build.go